requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.94.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2==2.9.5
pyasn1==0.4.8
pydantic==1.10.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.0
